requirements.txt
# Updated Requirements for OpenAI + Pinecone Integration
# Replace the existing requirements.txt with this updated version

# Core FastAPI and server
fastapi
uvicorn[standard]
python-multipart
websockets

# Async support and performance
aiohttp
uvloop

# Google Cloud APIs (STT and TTS)
google-cloud-speech>=2.21.0,<3.0.0
google-cloud-texttospeech>=2.14.0,<3.0.0
google-auth
google-api-core
google-cloud-core
grpcio
grpcio-status
protobuf
google-auth-oauthlib

# OpenAI and Pinecone (replacing Ollama and Chroma)
openai>=1.0.0
pinecone[grpc]>=3.0.0

# Audio Processing
librosa
soundfile
numpy
scipy
webrtcvad

# Document Processing
beautifulsoup4
pdfplumber
PyMuPDF
pypdf
python-docx
openpyxl
lxml

# Twilio Integration
twilio

# Data Handling and Utilities
pydantic
pydantic-settings
python-dotenv
requests
pandas
typing-extensions

# Caching and Session Management
redis

# Monitoring and Logging
structlog
prometheus-client

# Development and Testing
pytest
pytest-asyncio

# Language Processing
spacy
nltk
langdetect

# System and Performance Monitoring
psutil

# Authentication and Security
cryptography
pyjwt

# Additional Utilities
colorama
tqdm
click
pyaudio
Pillow

# Optimizations
ujson
orjson

# For memory optimization
memory-profiler